+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - Fixed a bug where using firework rocket in a block won't decrease the item count.
 - Fixed a client crash bug due to the incorrect absorption attribute sending.
 - Fixed a bug where water block cannot extinguish the fire on the entity.
+- Fixed a bug where the player dying in a non-overworld could not be properly teleported back to the spawn point in overworld.
 
 ### Removed